requirements.txt
streamlit
praw
pandas
matplotlib
textblob
python-dotenv
